+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Generic/PrettyCore.hs
@@ -1906,7 +1906,7 @@ ppShelleyDelegPredFailure :: ShelleyDelegPredFailure era -> PDoc
 ppShelleyDelegPredFailure x = case x of
   StakeKeyAlreadyRegisteredDELEG cred -> ppSexp "StakeKeyAlreadyRegisteredDELEG" [pcCredential cred]
   Shelley.StakeKeyNotRegisteredDELEG cred -> ppSexp "StakeKeyNotRegisteredDELEG" [pcCredential cred]
-  StakeKeyNonZeroAccountBalanceDELEG mcoin -> ppSexp " StakeKeyNonZeroAccountBalanceDELEG" [ppMaybe pcCoin mcoin]
+  StakeKeyNonZeroAccountBalanceDELEG c -> ppSexp "StakeKeyNonZeroAccountBalanceDELEG" [pcCoin c]
   StakeDelegationImpossibleDELEG cred -> ppSexp "StakeDelegationImpossibleDELEG" [pcCredential cred]
   WrongCertificateTypeDELEG -> ppSexp "WrongCertificateTypeDELEG" []
   GenesisKeyNotInMappingDELEG kh -> ppSexp "GenesisKeyNotInMappingDELEG" [pcKeyHash kh]